### 2022-03-08 #4233 @auroreloisy
Reviewer suggestion: vkrajnak

On Tue, Mar 8, 2022 at 10:37 AM The Open Journals editorial robot <
***@***.***> wrote:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04233/joss.04233/10.21105.joss.04233.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.04233/joss.04233/10.21105.joss.04233.pdf>
> 📄 👈
>
