### 2022-03-08 #4233 @auroreloisy
Suggested reviewers from the list: abhiramm7, g4brielvs, 11michalis11,
giodegas, ulf1, dglmoore

On Tue, Mar 8, 2022 at 4:30 PM The Open Journals editorial robot <
***@***.***> wrote:

> Assigned! @VivianePons <https://github.com/VivianePons> is now the editor
>
