### 2022-03-22 #4266 @editorialbot
[REVIEW]: OTTO: A Python package to simulate, solve and visualize the source-tracking POMDP
**Submitting author:** @auroreloisy (<a href="http://orcid.org/0000-0002-8089-8636">Aurore Loisy</a>)
**Repository:** https://github.com/C0PEP0D/otto
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
