### 2022-03-23 #4266 @auroreloisy
@VivianePons <https://github.com/VivianePons> Thank you for handling this
manuscript, should I provide you with additional reviewers suggestions?

On Tue, Mar 22, 2022 at 10:35 PM The Open Journals editorial robot <
***@***.***> wrote:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04266/joss.04266/10.21105.joss.04266.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.04266/joss.04266/10.21105.joss.04266.pdf>
> 📄 👈
>
