### 2022-04-04 C0PEP0D/otto#1 @11michalis11
JOSS submission review #4266
Overall the submission [(#4266)](https://github.com/openjournals/joss-reviews/issues/4266) seems very nice. Here's a list of minor comments:

- When running `visualise.py -i myparam` with `N_DIMS = 3` I get the following `ValueError` 👇 at the end of the run. All `.png` files are saved in the `frames` folder but the video is not created. The video is properly created for `N_DIMS = 1` and `N_DIMS = 2`. Maybe document this issue.
```
Traceback (most recent call last):
  File "/home/michalis/PhD/code/otto/otto/classes/visualization.py", line 423, in _draw_3D
    sc0 = ax[0].scatter(
  File "/home/michalis/anaconda3/envs/ottoenv/lib/python3.8/site-packages/matplotlib-3.5.1-py3.8-linux-
ValueError: s must be a scalar, or float array-like with the same size as x and y
```
- Consider using `python -m pytest tests` at the testing section of the `README.md` file instead of `pytest tests`. I couldn't run `pytest tests` because my terminal was looking at another version of `pytest`. Using the `-m` switch is generally good practice.
- Minor typo in the `Visualizing and evaluating a learned policy` section of the `README.md` file: I believe `YYmmdd-HHMMSS_value_model_bkp_i` should have been `YYmmdd-HHMMSS_model_bkp_i`
- Maybe include the contents of the `README.md` file in the documentation as well. The `README.md` file did a brilliant job on walking the user through the software but I feel like some of the sections should also be listed in the `docs` (Installation, How to use OTTO?).
- Line 8 of the draft paper: Not sure if it's a grammatical error or if I don't understand the sentence:
`Far from a “toy” POMDP, it incorporates physical models...`

In general this was easy to follow and it was a pleasure to review 👏👏
