+++ docs/quickstart.rst
@@ -1,0 +1,214 @@
+Quick start
+===========
+
+First steps
+-----------
+
+Go to the ``otto`` subdirectory.
+
+You will see that it is organized in **three main directories** corresponding to the **three main uses of OTTO**:
+
+  - ``evaluate``: for **evaluating the performance** of a policy
+  - ``learn``: for **learning a neural network policy** that solves the task
+  - ``visualize``: for **visualizing a search** episode
+
+The other directory, ``classes``, contains all the class definitions used by the main scripts.
+
+The three main directories share the same structure. They contain:
+
+  - ``*.py``: the main script
+  - ``parameters``: the directory to store input parameters
+  - ``outputs``: the directory to store files generated by the script (it will be created on first use)
+
+To use OTTO, go to the relevant main directory and run the corresponding script.
+For example, to visualize an episode, go to the ``visualize`` directory and run the ``visualize.py`` script with::
+
+    python3 visualize.py
+
+You should now see the rendering of a 1D search in a new window (it may be very short!).
+You can visualize another episode by using again the same command.
+
+Some logging information is displayed in the terminal as the script runs.
+In the rendering window, the first panel is a map of odor detections, and the second panel is the agent's belief
+(probability distribution over source locations).
+
+The videos have been saved as ``visualize/outputs/YYmmdd-HHMMSS_video.mp4`` where 'YYmmdd-HHMMSS' is a
+timestamp (the time you started the script).
+
+If you do not have FFmpeg or if you are using Windows, you will find instead frames saved
+in ``visualize/outputs/YYmmdd-HHMMSS_frames``.
+
+
+Changing parameters
+-------------------
+
+Many parameters (space dimension, domain size, source intensity, policy, ...) can be changed from the defaults.
+To run a script with different parameters, create a Python script that sets your parameters in the
+``parameters`` directory.
+
+A file ``myparam.py`` is already present in ``visualize/parameters/`` for this example.
+It contains a single line::
+
+    N_DIMS = 2
+
+which sets the dimensionality of the search to 2D (1D is the default).
+
+User-defined parameters are called by using the ``--input`` option followed by the name of the parameter file.
+For example, you can now visualize a search in 2D with::
+
+    python3 visualize.py --input myparam.py
+
+
+The ``--input`` option can be shortened to ``-i``, and the file name can be with or without ``.py``. So the command::
+
+    python3 visualize.py -i myparam
+
+will have the same effect.
+
+Each ``parameters`` directory contain sample parameter files called ``example*.py``.
+They show essential parameters you can play with, for example:
+
+  - ``N_DIMS`` sets the dimensionality of the search (1D, 2D, 3D), default is ``N_DIMS = 1``
+  - ``LAMBDA_OVER_DX`` controls the size the domain, default is ``LAMBDA_OVER_DX = 2.0``
+  - ``R_DT`` controls the source intensity, default is ``R_DT = 2.0``
+  - ``POLICY`` defines the policy to use, default is ``POLICY = 0`` (infotaxis)
+
+The definition of all parameters is provided :ref:`here<sec-visualize.py>`,
+and you can find their default values by examining the contents of ``__defaults.py``.
+
+Evaluating a policy
+-------------------
+
+The ``evaluate.py`` script (in the ``evaluate`` directory) computes many statistics that characterize the performance
+of a policy, such as
+
+  - probability of never finding the source,
+  - average time to find the source,
+  - probability distribution of arrival times,
+  - and much more.
+
+It does so essentially by running thousands of episodes in parallel and averaging over those.
+
+You can try with::
+
[…]
