### 2022-04-05 C0PEP0D/otto#1 @auroreloisy
We thank you for carefully evaluating our submission and for your positive feedback.
All the issues your have raised have been fixed in the last [commit](https://github.com/C0PEP0D/otto/commit/296109e10edc4f9f3de5c422a54fc2f388905917).
Thank your for your help with improving OTTO and the manuscript.
