### 2022-04-05 #4266 @auroreloisy
@11michalis11 Thank you for your positive feedback and for your help with improving our submission, we have replied in the issue.
@VivianePons We have addressed all the comments in the the last [commit](https://github.com/C0PEP0D/otto/commit/296109e10edc4f9f3de5c422a54fc2f388905917)
