### 2022-05-13 C0PEP0D/otto#2 @RomainAzais
JOSS submission review #4266 (2nd review)
Hi @auroreloisy
I just followed the procedure to install the software in a Conda environment.
The tests generate an error: `ModuleNotFoundError: No module named 'google.protobuf'`
I didn't have any problems with the installation (only some deprecation warnings) and Python 3.8.13 is installed in the env.
(NB: I'm on macOS Big Sur.)
Do you have any idea what's going on?
R.
Review thread: [#4266](https://github.com/openjournals/joss-reviews/issues/4266)
