### 2022-05-13 C0PEP0D/otto#2 @auroreloisy
Hi  @RomainAzais,
could you try to run the tests one by one to identify which one is broken?
Using
python3 -m pytest tests/test_defaults.py
python3 -m pytest tests/test_heuristicpolicy.py
etc
