### 2022-05-13 C0PEP0D/otto#2 @celoy
ça ressemble à un problème de tensorflow:

https://stackoverflow.com/questions/39582868/importerror-no-module-named-google-protobuf-but-its-definitely-installed <https://stackoverflow.com/questions/39582868/importerror-no-module-named-google-protobuf-but-its-definitely-installed>

> On May 13, 2022, at 12:19 PM, Aurore Loisy ***@***.***> wrote:
>
>
> Hi @RomainAzais <https://github.com/RomainAzais>,
> could you try to run the tests one by one to identify which one is broken?
> Using
> python3 -m pytest tests/test_defaults.py
> python3 -m pytest tests/test_heuristicpolicy.py
> etc
>
