### 2022-05-13 C0PEP0D/otto#2 @auroreloisy
(we validated the code with tensorflow 2.8.0)
