### 2022-05-13 C0PEP0D/otto#2 @RomainAzais
The error occurred in `test_rl.py` and `test_training.py`.
Actually, tensorflow was not installed in the environment even if you mentioned that the dependencies should be installed automatically. (And I said in my first message, no error at the installation.)
I tried several times through conda-forge (`conda install -c conda-forge tensorflow`) but it fails solving the environment.
