### 2022-05-13 ddd7242
updated requirements to fix issue #2
