+++ requirements.txt
@@ -1,8 +1,9 @@
 numpy >= 1.21
 scipy >= 1.7
 matplotlib >= 3.5
-tensorflow >= 2.3
-gym
+protobuf == 4.0.0rc2
+tensorflow == 2.8
+gym == 0.22
 sphinx >= 4.4
 sphinx-rtd-theme
 pytest
