### 2022-05-13 C0PEP0D/otto#2 @auroreloisy
We updated the requirements file to fix the issue, can to try to install otto from scratch again? I was able to reproduce your error, and it seems to work now.
