### 2022-05-19 C0PEP0D/otto#2 @RomainAzais
Hello!
I think your contribution is very nice. At this stage, I have two questions / comments:
- If I understand well, it is not possible for the user to customize the domain nor the starting location of the MDP and the one of the target (by the way, how are they selected?). I guess it drastically limits the applicability of your code. Am I right? If yes, do you plan to improve the software on this aspect? Maybe it is not your objective, but in that case, you could be more precise on what potential users can expect from it.
- I found the parameter files easy to use, but I was confused with the parameter `LAMBDA_OVER_DX` commented as "dimensionless problem size" in a parameter file and "controls the size the domain" ("of" is missing) in the ReadMe. One can easily understand its effect by testing different values, but do you think you could explain in one line its exact definition (or maybe refer to somewhere in a paper)? (The confusion might come from the name of the variable.)
R.
