### 2022-05-19 9b76ab8
edits about N_GRID issue #2
