### 2022-05-24 C0PEP0D/otto#2 @RomainAzais
Thank you @auroreloisy for your answer. Indeed, the documentation is clear on this subject and gives the right level of detail in my opinion (not only "the source-tracking POMDP" section, but also "policies" and "deep reinforcement learning" sections).
