### 2022-06-09 #4266 @RomainAzais
Hello,

My overall opinion of OTTO is very positive: the installation and testing procedures are clear, the readme file and the documentation are very complete and at the right level of detail, and the examples help to understand how to use it. I only encountered a small bug during the installation (with tensorflow apparently) but it has been solved. Setting up a ci/cd pipeline might avoid this kind of mishap.

My only (slightly) negative comment is related to the [discussion on the review thread](https://github.com/C0PEP0D/otto/issues/2#issuecomment-1131778327): reading only the paper, I expect (as a non-specialist) to be able to use OTTO on custom state spaces, where I can decide on target and tracker (starting) positions. (Perhaps this is a bias induced by the introduction of the paper on applications of this type of model.)

The [documentation](https://otto-c0pep0d.readthedocs.io/en/latest/pomdp.html) is precise on the model and strategies considered and helps to understand that this is not what is treated with this package. If I understood correctly, OTTO is more a tool for rigorous study and comparison of heuristic policies than a practical tool to be used on real-world examples. In my opinion, a sentence or two should be added in the paper to clarify the purpose of the package.

I still checked the "statement of need" box because this point is not blocking for me. On this subject, note that if we follow the checklist to the letter, a "statement of need" section should be present in the paper.

It was a pleasure to use and review OTTO. Thanks for the quick feedback during the review process, and sorry for my late feedback in this busy end of year period.

R.
