+++ paper/paper.bib
@@ -57,9 +57,11 @@
 @article{Loisy2022,
   title={Searching for a source without gradients: how good is infotaxis and how to beat it},
   author={Loisy, A. and Eloy, C.},
-  year={2021},
-  eprint={2112.10861},
-  archivePrefix={arXiv},
+  year={2022},
+  doi={10.1098/rspa.2022.0118},
+  journal={Proceedings of the Royal Society A},
+  volume = {to appear},
+  number = {xx--xx},
 }
 
 @article{Hutchinson2018,
