+++ paper/paper.md
@@ -46,6 +46,7 @@ The target audience of OTTO consists of researchers in biophysics, applied mathe
 
 `OTTO` (short for Odor-based Target Tracking Optimization) is a Python 
 package to learn, evaluate and visualize strategies for odor-based searches.
+It provides the means to rigorously study the statistical properties of any heuristic strategy, and to assess its performance with respect to that of a near-optimal solution.
 
 `OTTO` provides:
 
