### 2022-06-09 #4266 @auroreloisy
Dear @RomainAzais,
thank you very much for spending time on reviewing OTTO, and for your positive feedback.
Following your suggestion, we have added the following sentence in the paper to clarify the purpose of the package: "It is primarily designed to rigorously study the statistical properties of near-optimal strategies and of their heuristic approximations."
(cf [commit](https://github.com/C0PEP0D/otto/commit/9d4c4db3b5aa7759f281cf28a28e060e766905cc) and [commit](https://github.com/C0PEP0D/otto/commit/33af3a11bedeacbb24563406866ab167c83b847a)).
Regarding the "Statement of Need", it is actually present in the paper (first section).

Dear @VivianePons : we believe we have now addressed all issues raised by the reviewers.
