### 2022-06-10 #4266 @VivianePons
Good!

@RomainAzais can you confirm that you're satisfied with the answer from @auroreloisy to your comment?

@auroreloisy can you fix the Invalid DOI that is shown by the edit bot?
