### 2022-06-13 #4266 @RomainAzais
> Regarding the "Statement of Need", it is actually present in the paper (first section).

My mistake, I guess I was expecting it at the "summary" place. But that's perfectly fine.

> @RomainAzais can you confirm that you're satisfied with the answer from @auroreloisy to your comment?

Yes, I am.
