+++ paper/paper.bib
@@ -55,13 +55,14 @@
 }
 
 @article{Loisy2022,
-  title={Searching for a source without gradients: how good is infotaxis and how to beat it},
-  author={Loisy, A. and Eloy, C.},
-  year={2022},
-  doi={10.1098/rspa.2022.0118},
-  journal={Proceedings of the Royal Society A},
-  volume = {to appear},
-  number = {xx--xx},
+  title = {Searching for a Source without Gradients: How Good Is Infotaxis and How to Beat It},
+  author = {Loisy, A. and Eloy, C.},
+  year = {2022},
+  journal = {Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
+  volume = {478},
+  number = {2262},
+  pages = {20220118},
+  doi = {10.1098/rspa.2022.0118}
 }
 
 @article{Hutchinson2018,
