+++ docs/references.rst
@@ -3,10 +3,10 @@ References
 
 .. [Cassandra1996] Cassandra, A. R., Kaelbling, L. P., and Kurien, J. A. (1996). *Acting under uncertainty: Discrete
    Bayesian models for mobile-robot navigation*, Proceedings of IEEE/RSJ International Conference on Intelligent Robots
-   and Systems, Vol. 2, 963–972.
+   and Systems, Vol. 2, 963–972. `(doi:10.1109/IROS.1996.571080) <https://doi.org/10.1109/IROS.1996.571080>`_
 
 .. [Loisy2022] Loisy, A. and Eloy, C. (2022). *Searching for a source without gradients: how good is infotaxis and how
-   to beat it*, https://arxiv.org/abs/2112.10861
+   to beat it*, Proceedings of the Royal Society A **478**, 20220118. `(doi:10.1098/rspa.2022.0118) <https://doi.org/10.1098/rspa.2022.0118>`_
 
 .. [Vergassola2007] Vergassola, M., Villermaux, E., and Shraiman, B. (2007). *"Infotaxis" as a strategy for searching
-   without gradients*, Nature **445**, 406-409.
+   without gradients*, Nature **445**, 406-409. `(doi:10.1038/nature05464) <https://doi.org/10.1038/nature05464>`_
