### 2022-06-15 #4266 @VivianePons
Thank you @auroreloisy, I believe we have now everything and we will move to the last steps before final acceptance. Congrats! And thank you to @11michalis11 and @RomainAzais for the reviews.

@auroreloisy : can you make a tagged release of the software (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other). This needs to have the **same title and authors** as the submission. When done, send me the link and I will update the version and doi
